+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Terminate HTTP connection to S3 when response is closed (@janko-m)
+
 * Allow `Transfer-Encoding: chunked` to be used, meaning `Content-Length` can be blank (@janko-m)
 
 * Remove newlines from the base64-encoded CRC32 signature (@janko-m)